requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.7
cssselect2==0.8.0
Django==5.1.7
django-allauth==65.11.1
django-celery-beat==2.8.0
django-countries==7.6.1
django-filter==25.1
django-timezone-field==7.1
django-widget-tweaks==1.5.0
fonttools==4.58.4
idna==3.10
kombu==5.5.3
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
